refactor: Extract menu and forms from ui (Issue #3 Phase 2)
Extract menu and form rendering functions into dedicated modules
for better organization.

Changes:
- Created menu.rs (292 lines):
  * render_main_menu - Main category menu
  * render_tool_category - Tool list for category
  * render_tool_hotkeys - Keyboard shortcut display
- Created forms.rs (453 lines):
  * render_tool_form - Dynamic form builder
  * render_export_menu - Export format selection
  * render_filter_dialog - Packet filter configuration
- Updated core.rs with new module imports

File size reduction: 4,100 → 3,393 lines in core (17.2% reduction)
Cumulative: 5,020 → 3,393 lines (32.4% reduction)
Total: 5,797 lines across 6 modules

Testing:
✅ cargo check --package ubertooth-cli
✅ cargo clippy --package ubertooth-cli
